+++ CHANGELOG.md
@@ -1,7 +1,17 @@
 django-py3-ldap changelog
 =========================
 
-1.0.0
------
+1.0.1 - released on 2019-07-31
+------------------------------
+
+## added
+- management commands ldap_promote and ldap_sync_users.
+- Django test suite
+
+## fixed
+- setting variable names back to original format (`LDAP_*` -> `LDAP_AUTH_*`)
+
+1.0.0 - released on 2019-07-30
+------------------------------
 
 - First production release.